add /usr/local/sbin into path

+++ init_mac.sh
@@ -14,10 +14,18 @@ git config --global core.editor "vim"
 git config --global user.name "Xiaohan Zhang"
 git config --global user.email xiaohan.zhang@me.com
 
-mv ~/.vim vim.old
-mv ~/.vimrc vimrc.old
-mv ~/.bashrc bashrc.old
-mv ~/.bash_profile bash_profile.old
+if [ -f ~/.vim ]; then
+	mv ~/.vim vim.old
+fi
+if [ -f ~/.vimrc ]; then
+	mv ~/.vimrc vimrc.old
+fi
+if [ -f ~/.bashrc ]; then
+	mv ~/.bashrc bashrc.old
+fi
+if [ -f ~/.bash_profile ]; then
+	mv ~/.bash_profile bash_profile.old
+fi
 ln -sf ~/init/vim ~/.vim
 ln -sf ~/.vim/.vimrc ~/.vimrc
 ln -sf ~/init/bashrc_mac ~/.bashrc
